+++ apps/foreldrepengesoknad/src/steps/uttaksplan/hooks/useUttaksplanForEksisterendeSak.ts
@@ -4,13 +4,13 @@ import { ContextDataType, useContextGetData } from 'appData/FpDataContext';
 import dayjs from 'dayjs';
 import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
 import minMax from 'dayjs/plugin/minMax';
-import { useEffect, useRef } from 'react';
 
 import { ISO_DATE_FORMAT } from '@navikt/fp-constants';
-import { captureMessage, withScope } from '@navikt/fp-observability';
 import { UttakPeriodeAnnenpartEøs_fpoversikt, UttakPeriode_fpoversikt } from '@navikt/fp-types';
 import { Uttaksdagen } from '@navikt/fp-utils/src/uttak/Uttaksdagen';
 import { sorterUttakPerioder } from '@navikt/fp-uttaksplan';
+
+import { useLoggOverlappIVedtak } from './useLoggOverlappIVedtak';
 
 dayjs.extend(isSameOrBefore);
 dayjs.extend(minMax);
@@ -22,96 +22,20 @@ export const useUttaksplanForEksisterendeSak = (
 
     const sakerQuery = useQuery({ ...sakerOptions(), enabled: !!valgtEksisterendeSaksnr });
 
-    const harLoggetOverlapp = useRef(false);
-    useEffect(() => {
-        if (harLoggetOverlapp.current || !sakerQuery.data || !valgtEksisterendeSaksnr) {
-            return;
-        }
-        harLoggetOverlapp.current = true;
+    const valgtSak = sakerQuery.data?.foreldrepenger.find((sak) => sak.saksnummer === valgtEksisterendeSaksnr);
+    const perioderFraBackend = valgtSak?.gjeldendeVedtak?.perioder;
+    const justeringSøkerPerioder =
+        perioderFraBackend && perioderAnnenPart
+            ? midlertidigJusteringAvSamtidigUttak(perioderFraBackend, perioderAnnenPart)
+            : undefined;
 
-        const valgtSak = sakerQuery.data.foreldrepenger.find((sak) => sak.saksnummer === valgtEksisterendeSaksnr);
-        const perioderFraBackend = valgtSak?.gjeldendeVedtak?.perioder;
+    useLoggOverlappIVedtak(perioderFraBackend, perioderAnnenPart, justeringSøkerPerioder);
 
-        if (!perioderFraBackend) {
-            return;
-        }
-
-        const ugyldigeOverlappSøker = finnUgyldigeOverlapp(perioderFraBackend);
-        if (ugyldigeOverlappSøker.length > 0) {
-            withScope((scope) => {
-                scope.setLevel('warning');
-                scope.setTag('feiltype', 'uttaksplan-backend-overlapp');
-                scope.setExtra('kilde', 'søker');
-                scope.setExtra('antallUgyldigeOverlapp', ugyldigeOverlappSøker.length);
-                scope.setExtra(
-                    'ugyldigeOverlappPar',
-                    ugyldigeOverlappSøker.slice(0, 20).map(([a, b]) => ({
-                        a: periodeTilLoggObjekt(a),
-                        b: periodeTilLoggObjekt(b),
-                    })),
-                );
-                scope.setExtra('perioderFraBackend', perioderFraBackend.map(periodeTilLoggObjekt));
-                captureMessage('Eksisterande vedtak (søker) har ugyldig overlappande periodar', 'warning');
-            });
-        }
-
-        if (perioderAnnenPart && perioderAnnenPart.length > 0) {
-            const ugyldigeOverlappAnnenPart = finnUgyldigeOverlapp(perioderAnnenPart);
-            if (ugyldigeOverlappAnnenPart.length > 0) {
-                withScope((scope) => {
-                    scope.setLevel('warning');
-                    scope.setTag('feiltype', 'uttaksplan-backend-overlapp');
-                    scope.setExtra('kilde', 'annenPart');
-                    scope.setExtra('antallUgyldigeOverlapp', ugyldigeOverlappAnnenPart.length);
-                    scope.setExtra(
-                        'ugyldigeOverlappPar',
-                        ugyldigeOverlappAnnenPart.slice(0, 20).map(([a, b]) => ({
-                            a: periodeTilLoggObjekt(a),
-                            b: periodeTilLoggObjekt(b),
-                        })),
-                    );
-                    scope.setExtra('perioderFraBackend', perioderAnnenPart.map(periodeTilLoggObjekt));
-                    captureMessage('Eksisterande vedtak (annen part) har ugyldig overlappande periodar', 'warning');
-                });
-            }
-
-            const justert = midlertidigJusteringAvSamtidigUttak(perioderFraBackend, perioderAnnenPart);
-            const ugyldigeOverlappEtterJustering = finnUgyldigeOverlapp(justert);
-            if (ugyldigeOverlappEtterJustering.length > ugyldigeOverlappSøker.length) {
-                withScope((scope) => {
-                    scope.setLevel('warning');
-                    scope.setTag('feiltype', 'uttaksplan-midlertidig-justering-overlapp');
-                    scope.setExtra('antallFørJustering', ugyldigeOverlappSøker.length);
-                    scope.setExtra('antallEtterJustering', ugyldigeOverlappEtterJustering.length);
-                    scope.setExtra(
-                        'ugyldigeOverlappPar',
-                        ugyldigeOverlappEtterJustering.slice(0, 20).map(([a, b]) => ({
-                            a: periodeTilLoggObjekt(a),
-                            b: periodeTilLoggObjekt(b),
-                        })),
-                    );
-                    scope.setExtra('søkerPeriodeFørJustering', perioderFraBackend.map(periodeTilLoggObjekt));
-                    scope.setExtra('søkerPeriodeEtterJustering', justert.map(periodeTilLoggObjekt));
-                    scope.setExtra('annenPartPerioder', perioderAnnenPart.map(periodeTilLoggObjekt));
-                    captureMessage('midlertidigJusteringAvSamtidigUttak introduserte nye overlapp', 'warning');
-                });
-            }
-        }
-    }, [sakerQuery.data, valgtEksisterendeSaksnr, perioderAnnenPart]);
-
-    if (!sakerQuery?.data || !valgtEksisterendeSaksnr) {
+    if (!sakerQuery?.data || !valgtEksisterendeSaksnr || !valgtSak?.gjeldendeVedtak) {
         return undefined;
     }
 
-    const valgtSak = sakerQuery.data.foreldrepenger.find((sak) => sak.saksnummer === valgtEksisterendeSaksnr);
-
-    if (!valgtSak?.gjeldendeVedtak) {
-        return undefined;
-    }
-
-    const søkerPerioder = perioderAnnenPart
-        ? midlertidigJusteringAvSamtidigUttak(valgtSak.gjeldendeVedtak.perioder, perioderAnnenPart)
-        : valgtSak.gjeldendeVedtak.perioder;
+    const søkerPerioder = justeringSøkerPerioder ?? valgtSak.gjeldendeVedtak.perioder;
 
     const uttaksplan: Array<UttakPeriode_fpoversikt | UttakPeriodeAnnenpartEøs_fpoversikt> =
         fjernFrieUtsettelser(søkerPerioder);
@@ -204,41 +128,3 @@ const midlertidigJusteringAvSamtidigUttak = (
 
 const harOverlapp = (a: UttakPeriode_fpoversikt, b: UttakPeriode_fpoversikt) =>
     dayjs(a.fom).isSameOrBefore(b.tom, 'day') && dayjs(b.fom).isSameOrBefore(a.tom, 'day');
-
-const finnUgyldigeOverlapp = (
-    perioder: UttakPeriode_fpoversikt[],
-): Array<[UttakPeriode_fpoversikt, UttakPeriode_fpoversikt]> => {
-    const ugyldigeOverlapp: Array<[UttakPeriode_fpoversikt, UttakPeriode_fpoversikt]> = [];
-    for (let i = 0; i < perioder.length; i++) {
-        for (let j = i + 1; j < perioder.length; j++) {
-            const a = perioder[i]!;
-            const b = perioder[j]!;
-            if (
-                harOverlapp(a, b) &&
-                !(
-                    a.utsettelseÅrsak === undefined &&
-                    b.utsettelseÅrsak === undefined &&
-                    a.oppholdÅrsak === undefined &&
-                    b.oppholdÅrsak === undefined &&
-                    a.samtidigUttak !== undefined &&
-                    b.samtidigUttak !== undefined &&
-                    a.forelder !== b.forelder
-                )
-            ) {
-                ugyldigeOverlapp.push([a, b]);
-            }
-        }
-    }
-    return ugyldigeOverlapp;
-};
-
-const periodeTilLoggObjekt = (p: UttakPeriode_fpoversikt) => ({
-    fom: p.fom,
-    tom: p.tom,
-    forelder: p.forelder,
-    kontoType: p.kontoType,
-    utsettelseÅrsak: p.utsettelseÅrsak,
-    oppholdÅrsak: p.oppholdÅrsak,
-    overføringÅrsak: p.overføringÅrsak,
-    samtidigUttak: p.samtidigUttak,
-});
